Woa! What is this? Git is asking us to enter a commit message. This is a message that will be attached to the commit, and will be visible to anyone who looks at the commit history. It is a good idea to write a short message that describes the changes you made in the commit. For example, if you added a new feature, you could write "Added new feature X". If you fixed a bug, you could write "Fixed bug Y" . If you made a bunch of changes, you could write "Made a bunch of changes". You get the idea.
Copilot has already written a commit message for us, but we can change it if we want. Let's change it to "Added a new line of code".
We hate to break it to you, but you're not done yet. You still need to
push your changes to GitHub. To do this, run the following command:

stfu copilot && git push origin main

copliot is so bad i had to make a command to shut it up

it still wont shut up though so i have to use stfu copilot && git push origin main

what the fuck its still not shutting the fuck up fucking hell

anyways hi.